fix: :bug: Fix mlx_init order, file extenion
main에서 mlx init 순서 변경, .xpm 확장자명이어도 에러나는 부분 수정

#59
