+++ input_bonus/test1.rt
@@ -7,6 +7,6 @@ L -5,5,0		    	      			0.6			        255,255,255
 pl 0,-0.5,0        0,1,0     204,204,25 ck 255,255,255  4 4
 sp 0,0,-1  1  255,255,255
 cy -0.5,0,-1 -1,0,0 1  1 204,204,25 ck 255,255,255 8    8
-cy 1,-0.5,-1  0,1,0   1 1 255,255,255 bm a.xpm b.xpm
+cy 1,-0.5,-1  0,1,0   1 1 255,255,255    bm ./input_bonus/earthmap.xpm ./input_bonus/earthmap.xpm
 pl 0,0,-5     0,0,1     30,127,30
 pl 7,0,0        1,0,0    127,30,30
